Refactor dynamicconfig.FileBasedClient to make reusable
