+++ common/dynamicconfig/client.go
@@ -1,6 +1,8 @@
 package dynamicconfig
 
 import (
+	"strings"
+
 	enumspb "go.temporal.io/api/enums/v1"
 	enumsspb "go.temporal.io/server/api/enums/v1"
 )
@@ -31,7 +33,7 @@ type (
 	NotifyingClient interface {
 		// Adds a subscription to all updates from this Client. `update` will be called on any
 		// change to the current value set. The caller should call `cancel` to cancel the
-		// subscription. Calls to `update` will not be made concurrently.
+		// subscription.
 		Subscribe(update ClientUpdateFunc) (cancel func())
 	}
 
@@ -96,4 +98,8 @@ type (
 
 func (k Key) String() string {
 	return string(k)
+}
+
+func (k Key) Lower() Key {
+	return Key(strings.ToLower(string(k)))
 }
