+++ common/dynamicconfig/registry.go
@@ -2,13 +2,12 @@ package dynamicconfig
 
 import (
 	"fmt"
-	"strings"
 	"sync/atomic"
 )
 
 type (
 	registry struct {
-		settings map[string]GenericSetting
+		settings map[Key]GenericSetting
 		queried  atomic.Bool
 	}
 )
@@ -22,20 +21,20 @@ func register(s GenericSetting) {
 		panic("dynamicconfig.New*Setting must only be called from static initializers")
 	}
 	if globalRegistry.settings == nil {
-		globalRegistry.settings = make(map[string]GenericSetting)
+		globalRegistry.settings = make(map[Key]GenericSetting)
 	}
-	keyStr := strings.ToLower(s.Key().String())
-	if globalRegistry.settings[keyStr] != nil {
-		panic(fmt.Sprintf("duplicate registration of dynamic config key: %q", keyStr))
+	lowerKey := s.Key().Lower()
+	if globalRegistry.settings[lowerKey] != nil {
+		panic(fmt.Sprintf("duplicate registration of dynamic config key: %q", lowerKey))
 	}
-	globalRegistry.settings[keyStr] = s
+	globalRegistry.settings[lowerKey] = s
 }
 
 func queryRegistry(k Key) GenericSetting {
 	if !globalRegistry.queried.Load() {
 		globalRegistry.queried.Store(true)
 	}
-	return globalRegistry.settings[strings.ToLower(k.String())]
+	return globalRegistry.settings[k.Lower()]
 }
 
 // For testing only; do not call from regular code!
